requirements.txt
config==0.3.9
dj-database-url==0.5.0
Django==2.2.2
django-bootstrap4==0.0.8
django-crispy-forms==1.7.2
django-filter==2.1.0
django-heroku==0.3.1
djangorestframework==3.9.4
gunicorn==19.9.0
Pillow==6.0.0
psycopg2==2.8.2
python-decouple==3.1
pytz==2019.1
sqlparse==0.3.0
whitenoise==4.1.2
django-daterangefilter==1.0.0

celery
django-celery==3.1.15
redis

django_celery_beat